[S12P21A402-156] feat: Implement repository and refactor ddl

+++ server/src/main/java/com/lufin/server/stock/dto/StockPortfolioResponseDto.java
@@ -10,7 +10,7 @@ public class StockPortfolioResponseDto {
 		Integer quantity,
 		Integer totalPurchaseAmount,
 		Integer totalSellAmount,
-		Integer totalReturn,
+		Double totalReturn,
 		Double totalReturnRate,
 		LocalDateTime createdAt,
 		LocalDateTime updatedAt
